migrate device state to device data
